# packages/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pure Botanica API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AssessmentInputGender = typeof AssessmentInputGender[keyof typeof AssessmentInputGender];


export const AssessmentInputGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export type AssessmentInputActivityLevel = typeof AssessmentInputActivityLevel[keyof typeof AssessmentInputActivityLevel];


export const AssessmentInputActivityLevel = {
  calm: 'calm',
  average: 'average',
  very_active: 'very_active',
} as const;

export type AssessmentInputFocusDifficulty = typeof AssessmentInputFocusDifficulty[keyof typeof AssessmentInputFocusDifficulty];


export const AssessmentInputFocusDifficulty = {
  never: 'never',
  sometimes: 'sometimes',
  often: 'often',
} as const;

export type AssessmentInputHyperactivity = typeof AssessmentInputHyperactivity[keyof typeof AssessmentInputHyperactivity];


export const AssessmentInputHyperactivity = {
  never: 'never',
  sometimes: 'sometimes',
  often: 'often',
} as const;

export type AssessmentInputHomework = typeof AssessmentInputHomework[keyof typeof AssessmentInputHomework];


export const AssessmentInputHomework = {
  yes: 'yes',
  sometimes: 'sometimes',
  no: 'no',
} as const;

export type AssessmentInputDiet = typeof AssessmentInputDiet[keyof typeof AssessmentInputDiet];


export const AssessmentInputDiet = {
  excellent: 'excellent',
  good: 'good',
  average: 'average',
  poor: 'poor',
} as const;

export type AssessmentInputVegetables = typeof AssessmentInputVegetables[keyof typeof AssessmentInputVegetables];


export const AssessmentInputVegetables = {
  NUMBER_0: '0',
  NUMBER_1: '1',
  '2+': '2+',
  '3+': '3+',
} as const;

export type AssessmentInputSupplements = typeof AssessmentInputSupplements[keyof typeof AssessmentInputSupplements];


export const AssessmentInputSupplements = {
  yes: 'yes',
  no: 'no',
} as const;

export interface AssessmentInput {
  parentName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  childName: string;
  /**
     * @minimum 3
     * @maximum 12
     */
  age: number;
  gender: AssessmentInputGender;
  activityLevel: AssessmentInputActivityLevel;
  focusDifficulty: AssessmentInputFocusDifficulty;
  hyperactivity: AssessmentInputHyperactivity;
  homework: AssessmentInputHomework;
  diet: AssessmentInputDiet;
  vegetables: AssessmentInputVegetables;
  supplements: AssessmentInputSupplements;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  medications?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type AssessmentResultRecommendation = typeof AssessmentResultRecommendation[keyof typeof AssessmentResultRecommendation];


export const AssessmentResultRecommendation = {
  focus: 'focus',
  mineral: 'mineral',
  both: 'both',
} as const;

export interface AssessmentResult {
  id: number;
  parentName: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  childName: string;
  age: number;
  gender: string;
  activityLevel: string;
  focusDifficulty: string;
  hyperactivity: string;
  homework: string;
  diet: string;
  vegetables: string;
  supplements: string;
  /** @nullable */
  allergies?: string | null;
  /** @nullable */
  medications?: string | null;
  /** @nullable */
  notes?: string | null;
  focusScore: number;
  mineralScore: number;
  recommendation: AssessmentResultRecommendation;
  createdAt: string;
}

export interface AssessmentListItem {
  id: number;
  parentName: string;
  email: string;
  childName: string;
  age: number;
  gender: string;
  recommendation: string;
  focusScore: number;
  mineralScore: number;
  createdAt: string;
}

export interface AssessmentListResponse {
  data: AssessmentListItem[];
  total: number;
  page: number;
  limit: number;
}

export interface AdminCredentials {
  password: string;
}

export interface AdminMeResponse {
  authenticated: boolean;
}

export interface AgeDistItem {
  age: number;
  count: number;
}

export interface GenderDistItem {
  gender: string;
  count: number;
}

export interface RecommendationDistItem {
  recommendation: string;
  count: number;
}

export interface SymptomItem {
  symptom: string;
  count: number;
}

export interface AdminStats {
  totalAssessments: number;
  focusCount: number;
  mineralCount: number;
  bothCount: number;
  ageDistribution: AgeDistItem[];
  genderDistribution: GenderDistItem[];
  recommendationDistribution: RecommendationDistItem[];
  commonSymptoms: SymptomItem[];
}

export type ListAdminAssessmentsParams = {
search?: string;
recommendation?: string;
page?: number;
limit?: number;
};