# packages/api-zod/src/generated/types/adminCredentials.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pure Botanica API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminCredentials {
  password: string;
}