# packages/api-zod/src/generated/types/assessmentListItem.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Pure Botanica API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AssessmentListItem {
  id: number;
  parentName: string;
  email: string;
  childName: string;
  age: number;
  gender: string;
  recommendation: string;
  focusScore: number;
  mineralScore: number;
  createdAt: Date;
}